Day 5 Part 2
 I started thinking the regex approach would be good. It wasn't.
This slowed me down significantly on the first part as I struggled
to get the pattern to match. Switching it to a simple string search
was a good move.
 I then ruined myself by thinking that a simple swap would suffice
in the real input as it did in the example. Of course this was
untrue so a refactor to interate over rules until the entire update
was corrected cost me time.